requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
pyyaml>=6.0.1
colorlog>=6.8.0
schedule>=1.2.0
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
